requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
jinja2==3.1.3
python-dotenv==1.0.1
fastapi-mail==1.4.1
requests==2.31.0
gunicorn==20.1.0
python-multipart==0.0.9
psycopg2-binary==2.9.9

